:recycle: backend refactor - created apps for handling games and actions, separately - now we have the following apps: users, games, actions and api - the api app now only serves as a meeting point for all the apps
