+++ backend/actions/serializers.py
@@ -1,22 +1,16 @@
-from .models import Game
 from rest_framework import serializers
-from .models import Game, Ratings
-
-
-class GameSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Game
-        fields = ('igdb', 'name',)
-
+from .models import Ratings
+        
 
 class ActionSerializer(serializers.Serializer):
     game = serializers.IntegerField()
     user = serializers.IntegerField()
     action = serializers.CharField(max_length=16)
     value = serializers.BooleanField()
-
+    
 
 class RatingSerializer(serializers.ModelSerializer):
     class Meta:
         model = Ratings
         fields = ('game', 'user', 'rating')
+
