+++ backend/backend/settings.py
@@ -45,7 +45,9 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'django.contrib.sites',
+    'actions',
     'api',
+    'games',
     'users',
     'corsheaders',
     'rest_framework',
