+++ backend/games/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Game
+
+
+class GameSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Game
+        fields = ('igdb', 'name',)
